backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
httpx>=0.24.0
python-dotenv>=1.0.0
